requirements.txt
streamlit==1.35.0
langchain==0.2.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
PyMuPDF==1.23.22
python-docx==1.1.0
python-pptx==0.6.23
transformers==4.41.1
torch==2.2.2
